[Build] Fixes and Raw Pointer refactoring - Fix language build (namespace messages) - Requires C++14 for std::unique_ptr creation - Replace pointers with smart pointers

+++ src/interface/keybdinput.cpp
@@ -259,7 +259,7 @@ void CAKeybdInput::midiInEventToScore(CAScoreView *v, QVector<unsigned char> m) 
 
 		// scene tracking
 		QRectF scene = v->worldCoords();
-		int xlast = v->timeToCoordsSimpleVersion( voice->lastTimeStart() );
+		double xlast = v->timeToCoordsSimpleVersion( voice->lastTimeStart() );
 		if ( ((xlast+50) > scene.right()) ) {	// the magic number 50 should be defined, ist the width of an element
 			scene.translate( scene.width()/2, 0 );
 			v->setWorldCoords(scene, false, true );
